Move the "application" code to package `app`
